some test code, html bug fixes and auth.md updates

+++ backend/internal/config/constants.go
@@ -6,8 +6,8 @@ import "time"
 const (
 	ServiceName = "Impstagram"
 
-	MinAge uint8 = 18
-	MaxAge uint8 = 120
+	MinAge uint16 = 18
+	MaxAge uint16 = 120
 
 	EmailID string = Env + "@impstagram.ripgod.xyz"
 )
